requirements.txt
SQLAlchemy==1.3.20
uvicorn==0.13.3
pydantic==1.7.3
FastAPI_SQLAlchemy==0.2.1
twilio==6.52.0
fastapi==0.63.0
fastapi_utils==0.2.1
mysql_connector_repackaged==0.3.1
python-dotenv==0.15.0
